Migrated the sugarfunge-metadata.scale to the sugarfunge-api-types
