+++ crates/sugarfunge-api-types/Cargo.toml
@@ -6,5 +6,41 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+env_logger = "0.8"
+structopt = "0.3"
+rand = "0.8.3"
+hex = "0.4.3"
+actix-web = "4.0.1"
+actix-cors = "0.6.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+jsonrpsee-types = "0.5.1"
+url = { version = "2", features = ["serde"] }
+scale-info = { version = "1.0.0", default-features = false, features = [] }
+codec = { package = "parity-scale-codec", version = "2", default-features = false, features = ["derive", "full"] }
+derive_more = { version = "0.99.0", default-features = false }
+
+[dependencies.subxt]
+# path = "../subxt"
+git = "https://github.com/SugarFunge/subxt.git"
+rev = "15b4f70c40d72ceb259a63bf8c62fb3022e29a86"
+
+[dependencies.sp-keyring]
+git = "https://github.com/paritytech/substrate.git"
+branch = "polkadot-v0.9.17"
+package = "sp-keyring"
+
+[dependencies.sp-std]
+git = "https://github.com/paritytech/substrate.git"
+branch = "polkadot-v0.9.17"
+package = "sp-std"
+
+[dependencies.sp-core]
+git = "https://github.com/paritytech/substrate.git"
+branch = "polkadot-v0.9.17"
+package = "sp-core"
+features = ["full_crypto"]
+
+[features]
+default = ["std"]
+std = ["codec/std", "scale-info/std"]
