+++ crates/sugarfunge-api-types/sugarfunge_metadata.scale
