+++ src/escrow.rs
@@ -1,11 +1,11 @@
 use crate::state::*;
-use crate::sugarfunge;
 use crate::util::*;
 use actix_web::{error, web, HttpResponse};
 use serde_json::json;
 use std::str::FromStr;
 use subxt::PairSigner;
 use sugarfunge_api_types::escrow_types::*;
+use sugarfunge_api_types::sugarfunge;
 
 pub async fn create_escrow(
     data: web::Data<AppState>,
